requirements.txt
Flask
flask_mysqldb
flask-hashing
flask-cors
pyjwt
pymysql
python-dotenv
tensorflow==2.11.0
numpy==1.23.5
librosa==0.10.0
resampy==0.4.2
flask-cors